Split test_base.py into individual feature tests
New files:

- test_app_auth.py
- test_app_config.py
- test_app_firestore_service.py
- test_main.py
